pyproject.toml
[project]
name = "agent-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "langchain-groq>=0.3.2",
    "langgraph>=0.3.31",
    "python-dotenv>=1.1.0",
]
